requirements.txt
numpy>=1.23
pandas>=2.0
scikit-learn>=1.3
scipy>=1.10
pyyaml>=6.0
joblib>=1.3
pandera>=0.20
matplotlib>=3.7
structlog>=24.0
tabulate